Refresh Token implementation
Refresh token will be used to generate a new access token when the access token is expired or not valid for whatever reason. Also added a custom logout handler which will clear the SecurityContextHolder when called. Currently storing the access tokens in the db, but might implement storing the refresh-token instead. Will have to dig deeper about which implementation is better. CIAO RAGGAZZI.